+++ x_render.c
@@ -6,15 +6,15 @@
 /*   By: adavis <adavis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/31 20:47:32 by adavis            #+#    #+#             */
-/*   Updated: 2019/09/02 18:14:55 by adavis           ###   ########.fr       */
+/*   Updated: 2019/09/03 12:47:52 by adavis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int		x_nbrlen(int nbr)
+size_t	x_nbrlen(long long nbr)
 {
-	int		cnt;
+	size_t	cnt;
 
 	cnt = 1;
 	while (nbr /= 16)
@@ -22,7 +22,7 @@ int		x_nbrlen(int nbr)
 	return (cnt);
 }
 
-size_t	x_count_len(int d, t_params *params)
+size_t	x_count_len(long long d, t_params *params)
 {
 	size_t	len;
 	int		tmp;
@@ -33,12 +33,12 @@ size_t	x_count_len(int d, t_params *params)
 		len++;
 	if (params->alternate)
 		len += 2;
-	if (params->width > len)
+	if ((int)params->width > len)
 		len += params->width - len;
 	return (len);
 }
 
-void	x_render_left(int d, t_params *params, t_bool upper)
+void	x_render_left(long long d, t_params *params, t_bool upper)
 {
 	if (params->alternate)
 		params->width--;
@@ -47,26 +47,26 @@ void	x_render_left(int d, t_params *params, t_bool upper)
 		if (params->alternate)
 			ft_putstr(upper ? "0X" : "0x");
 		ft_putnbr_base(d, 16, upper);
-		while ((params->width--) - x_nbrlen(d) > 0)
+		while ((int)(params->width--) - (int)x_nbrlen(d) > 0)
 			ft_putchar(' ');
 	}
 	else
 		ft_putnbr_base(d, 16, upper);
 }
 
-void	x_render_right(int d, t_params *params, t_bool upper)
+void	x_render_right(long long d, t_params *params, t_bool upper)
 {
 	if (params->zeros)
 	{
 		if (params->alternate)
 			ft_putstr(upper ? "0X" : "0x");
-		while ((params->width--) - x_nbrlen(d) - params->alternate * 2 > 0)
+		while ((int)(params->width--) - (int)x_nbrlen(d) - params->alternate * 2 > 0)
 			ft_putchar(params->zeros ? '0' : ' ');
 		ft_putnbr_base(d, 16, upper);
 	}
 	else
 	{
-		while ((params->width--) - x_nbrlen(d) - params->alternate * 2 > 0)
+		while ((int)(params->width--) - (int)x_nbrlen(d) - params->alternate * 2 > 0)
 			ft_putchar(' ');
 		if (params->alternate)
 			ft_putstr(upper ? "0X" : "0x");
@@ -74,7 +74,7 @@ void	x_render_right(int d, t_params *params, t_bool upper)
 	}
 }
 
-int		x_render(int d, t_params *params, t_bool upper)
+int		x_render(long long d, t_params *params, t_bool upper)
 {
 	size_t	len;
